requirements.txt
requests>=2.28.0
anthropic>=0.5.0
python-dotenv>=1.0.0
openai>=1.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
sseclient-py>=1.7.2 